Move import-export utils
