+++ src/ggrc-client/js/components/import-export/download-template/download-template_spec.js
@@ -5,6 +5,7 @@
 
 import {getComponentVM} from '../../../../js_specs/spec_helpers';
 import Component from './download-template';
+import * as Utils from '../import-export-utils';
 
 describe('download-template component', () => {
   let vm;
@@ -46,14 +47,14 @@ describe('download-template component', () => {
 
   describe('downloadCSV method', () => {
     it('should do ajax call in proper format', (done) => {
-      spyOn(GGRC.Utils, 'export_request')
+      spyOn(Utils, 'exportRequest')
         .and.returnValue(can.Deferred().resolve('FooBar'));
-      spyOn(GGRC.Utils, 'download');
+      spyOn(Utils, 'download');
 
       vm.attr('selected', [{name: 'Foo'}, {name: 'Bar'}, {name: 'Baz'}]);
 
       vm.downloadCSV().then(() => {
-        expect(GGRC.Utils.export_request)
+        expect(Utils.exportRequest)
           .toHaveBeenCalledWith({
             data: {
               objects: [
@@ -64,7 +65,7 @@ describe('download-template component', () => {
               export_to: 'csv',
             },
           });
-        expect(GGRC.Utils.download)
+        expect(Utils.download)
           .toHaveBeenCalledWith('import_template.csv', 'FooBar');
         expect(vm.attr('modalState.open')).toEqual(false);
 
@@ -73,28 +74,28 @@ describe('download-template component', () => {
     });
 
     it('should not do ajax call if nothing selected', () => {
-      spyOn(GGRC.Utils, 'export_request');
-      spyOn(GGRC.Utils, 'download');
+      spyOn(Utils, 'exportRequest');
+      spyOn(Utils, 'download');
 
       vm.attr('selected', []);
 
       vm.downloadCSV();
 
-      expect(GGRC.Utils.export_request).not.toHaveBeenCalled();
-      expect(GGRC.Utils.download).not.toHaveBeenCalled();
+      expect(Utils.exportRequest).not.toHaveBeenCalled();
+      expect(Utils.download).not.toHaveBeenCalled();
     });
 
     it('should not download CSV and should close the modal after error during' +
-      '"export_request" phase', (done) => {
-      spyOn(GGRC.Utils, 'export_request')
+      '"exportRequest" phase', (done) => {
+      spyOn(Utils, 'exportRequest')
         .and.returnValue(can.Deferred().reject());
-      spyOn(GGRC.Utils, 'download');
+      spyOn(Utils, 'download');
 
       vm.attr('selected', [1, 2, 3]);
 
       vm.downloadCSV().fail(() => {
-        expect(GGRC.Utils.export_request).toHaveBeenCalled();
-        expect(GGRC.Utils.download).not.toHaveBeenCalled();
+        expect(Utils.exportRequest).toHaveBeenCalled();
+        expect(Utils.download).not.toHaveBeenCalled();
         expect(vm.attr('modalState.open')).toEqual(false);
 
         done();
